feat(builtin): Use help tags information for builtin commands
This was a semi manual proccess, and should use a script instead.
Currently, their is no way to check if the ex-command requires an
argument.
This should be done to prevent the need on hitting enter on some
commands such as copen, bnext, etc
Some ex-commands are also not parsed correctly, since the help page for
them is inconsitent